+++ stencil-workspace/src/interfaces.d.ts
@@ -3,3 +3,4 @@ export { Crumb } from './components/modus-breadcrumb/modus-breadcrumb';
 export { ModusNavbarApp } from './components/modus-navbar/apps-menu/modus-navbar-apps-menu';
 export { ModusNavbarProfileMenuLink } from './components/modus-navbar/profile-menu/modus-navbar-profile-menu';
 export * from './components/modus-data-table/modus-data-table.models';
+export { ModusAutocompleteOption } from './components/modus-autocomplete/modus-autocomplete';