+++ stencil-workspace/src/components/modus-navbar/apps-menu/modus-navbar-apps-menu.tsx
@@ -1,5 +1,5 @@
 // eslint-disable-next-line
-import { Component, Prop, h } from '@stencil/core';
+import { Component, Prop, h, EventEmitter, Event } from '@stencil/core';
 
 export interface ModusNavbarApp {
   description?: string;
@@ -19,8 +19,11 @@ export class ModusNavbarAppsMenu {
   @Prop() apps: ModusNavbarApp[];
   @Prop() reverse: boolean;
 
+  @Event() appOpen: EventEmitter<ModusNavbarApp>;
+
   clickAppHandler(app: ModusNavbarApp): void {
     window.open(app.url, '_blank');
+    this.appOpen.emit(app);
   }
 
   render(): unknown {