+++ CHANGELOG.md
@@ -32,6 +32,8 @@ View all releases at: <https://github.com/trimble-oss/modus-web-components/relea
 - Added password text toggle feature to `modus-text-input` component.
 - Modus Accordions now have transition animations
 
+- Add `modus-autocomplete`
+
 ## 0.1.28 - 2022-12-06
 
 ### Fixed